Add files via upload
In this lab, I set up a basic home SOC in Azure from scratch. Using a free Azure subscription, I walk through creating a virtual machine (VM), opening it to the internet as a honeypot, and forwarding logs to a central repository. I then integrate Microsoft Sentinel to analyse real-world attack data. 
In this lab I cover the following: 
•	Creating an Azure subscription and setting up a VM 
•	Configuring Log Analytics Workspace 
•	Forwarding logs and integrating with Sentinel 
•	Querying failed login attempts and visualizing attack sources 
•	Building an attack map to track real-time hacker activity 
This project is great for cybersecurity personnel, like myself, looking to practice log analysis, threat detection, and SOC operations in a real-world cloud environment.